Migrate media picker into separate property editors for single and range.
