+++ src/Umbraco.Core/PropertyEditors/MediaPicker3Configuration.cs
@@ -3,17 +3,12 @@ namespace Umbraco.Cms.Core.PropertyEditors;
 /// <summary>
 ///     Represents the configuration for the media picker value editor.
 /// </summary>
-public class MediaPicker3Configuration : IIgnoreUserStartNodesConfig
+public class MediaPicker3Configuration : MediaPickerConfigurationBase
 {
-    /// <summary>
-    /// Gets or sets the media type filter.
-    /// </summary>
-    [ConfigurationField("filter")]
-    public string? Filter { get; set; }
-
     /// <summary>
     /// Gets or sets a value indicating whether multiple media items can be selected.
     /// </summary>
+    [Obsolete("A media picker holding a single item is now its own property editor. Scheduled for removal in Umbraco 21.")]
     [ConfigurationField("multiple")]
     public bool Multiple { get; set; }
 
@@ -22,28 +17,6 @@ public class MediaPicker3Configuration : IIgnoreUserStartNodesConfig
     /// </summary>
     [ConfigurationField("validationLimit")]
     public NumberRange ValidationLimit { get; set; } = new();
-
-    /// <summary>
-    /// Gets or sets the start node ID for the media picker.
-    /// </summary>
-    [ConfigurationField("startNodeId")]
-    public Guid? StartNodeId { get; set; }
-
-    /// <summary>
-    /// Gets or sets a value indicating whether local focal point editing is enabled.
-    /// </summary>
-    [ConfigurationField("enableLocalFocalPoint")]
-    public bool EnableLocalFocalPoint { get; set; }
-
-    /// <summary>
-    /// Gets or sets the configured image crops.
-    /// </summary>
-    [ConfigurationField("crops")]
-    public CropConfiguration[]? Crops { get; set; }
-
-    /// <inheritdoc />
-    [ConfigurationField(Constants.DataTypes.ReservedPreValueKeys.IgnoreUserStartNodes)]
-    public bool IgnoreUserStartNodes { get; set; }
 
     /// <summary>
     /// Represents a numeric range with optional minimum and maximum values.
@@ -59,26 +32,5 @@ public class MediaPicker3Configuration : IIgnoreUserStartNodesConfig
         /// Gets or sets the maximum value of the range.
         /// </summary>
         public int? Max { get; set; }
-    }
-
-    /// <summary>
-    /// Represents an image crop configuration.
-    /// </summary>
-    public class CropConfiguration
-    {
-        /// <summary>
-        /// Gets or sets the alias of the crop.
-        /// </summary>
-        public string? Alias { get; set; }
-
-        /// <summary>
-        /// Gets or sets the width of the crop in pixels.
-        /// </summary>
-        public int Width { get; set; }
-
-        /// <summary>
-        /// Gets or sets the height of the crop in pixels.
-        /// </summary>
-        public int Height { get; set; }
     }
 }
