+++ src/Umbraco.Core/PropertyEditors/SingleMediaPickerConfiguration.cs
@@ -1,0 +1,8 @@
+namespace Umbraco.Cms.Core.PropertyEditors;
+
+/// <summary>
+///     Represents the configuration for the single media picker value editor.
+/// </summary>
+public class SingleMediaPickerConfiguration : MediaPickerConfigurationBase
+{
+}
