+++ src/Umbraco.Infrastructure/PropertyEditors/ValueConverters/SingleMediaPickerValueConverter.cs
@@ -8,21 +8,21 @@ using Umbraco.Cms.Infrastructure.DeliveryApi;
 namespace Umbraco.Cms.Core.PropertyEditors.ValueConverters;
 
 /// <summary>
-/// Converts the value stored by the Media Picker with Crops property editor into a strongly-typed object
-/// representing the selected media items and their associated crop data, making it accessible for use in code.
+/// Converts the value stored by the single media picker property editor into a strongly-typed object representing
+/// the selected media item and its associated crop data, making it accessible for use in code.
 /// </summary>
 [DefaultPropertyValueConverter]
-public class MediaPickerWithCropsValueConverter : MediaPickerWithCropsValueConverterBase
+public class SingleMediaPickerValueConverter : MediaPickerWithCropsValueConverterBase
 {
     /// <summary>
-    /// Initializes a new instance of the <see cref="MediaPickerWithCropsValueConverter"/> class.
+    /// Initializes a new instance of the <see cref="SingleMediaPickerValueConverter"/> class.
     /// </summary>
     /// <param name="publishedMediaCache">Provides access to the cache of published media items.</param>
     /// <param name="publishedUrlProvider">Resolves URLs for published content and media.</param>
     /// <param name="publishedValueFallback">Handles fallback logic for published property values.</param>
     /// <param name="jsonSerializer">Serializes and deserializes JSON data for property values.</param>
     /// <param name="apiMediaWithCropsBuilder">Builds API representations of media items with crop data.</param>
-    public MediaPickerWithCropsValueConverter(
+    public SingleMediaPickerValueConverter(
         IPublishedMediaCache publishedMediaCache,
         IPublishedUrlProvider publishedUrlProvider,
         IPublishedValueFallback publishedValueFallback,
@@ -33,7 +33,7 @@ public class MediaPickerWithCropsValueConverter : MediaPickerWithCropsValueConve
     }
 
     /// <inheritdoc />
-    protected override bool HoldsMultipleItems => true;
+    protected override bool HoldsMultipleItems => false;
 
     /// <summary>
     /// Determines whether this converter applies to the specified property type.
@@ -41,13 +41,13 @@ public class MediaPickerWithCropsValueConverter : MediaPickerWithCropsValueConve
     /// <param name="propertyType">The property type to check.</param>
     /// <returns><c>true</c> if this converter is applicable; otherwise, <c>false</c>.</returns>
     public override bool IsConverter(IPublishedPropertyType propertyType) =>
-        propertyType.EditorAlias.Equals(Constants.PropertyEditors.Aliases.MediaPicker3);
+        propertyType.EditorAlias.Equals(Constants.PropertyEditors.Aliases.SingleMediaPicker);
 
     /// <summary>
     /// Determines the CLR type returned for the property value.
     /// </summary>
     /// <param name="propertyType">The published property type to inspect.</param>
-    /// <returns><see cref="IEnumerable{MediaWithCrops}" />.</returns>
+    /// <returns><see cref="MediaWithCrops" />.</returns>
     public override Type GetPropertyValueType(IPublishedPropertyType propertyType)
-        => typeof(IEnumerable<MediaWithCrops>);
+        => typeof(MediaWithCrops);
 }
