+++ src/Umbraco.Web.UI.Client/src/packages/media/media/property-editors/media-picker/index.ts
@@ -1,1 +1,2 @@
-export * from './property-editor-ui-media-picker.element.js';
+export { UmbPropertyEditorUIMediaPickerElement } from './property-editor-ui-media-picker.element.js';
+export { UmbPropertyEditorUISingleMediaPickerElement } from './property-editor-ui-single-media-picker.element.js';
