+++ src/Umbraco.Web.UI.Client/src/packages/media/media/property-editors/media-picker/manifests.ts
@@ -1,5 +1,23 @@
 import { manifest as schemaManifest } from './Umbraco.MediaPicker.js';
+import { manifest as singleSchemaManifest } from './Umbraco.SingleMediaPicker.js';
 import { manifests as valueSummaryManifests } from './value-summary/manifests.js';
+
+const keywords = [
+	'select',
+	'image',
+	'photo',
+	'picture',
+	'banner',
+	'thumbnail',
+	'logo',
+	'avatar',
+	'gallery',
+	'media',
+	'video',
+	'file',
+	'attachment',
+	'cover',
+];
 
 export const manifests: Array<UmbExtensionManifest> = [
 	{
@@ -10,27 +28,27 @@ export const manifests: Array<UmbExtensionManifest> = [
 		meta: {
 			label: 'Media Picker',
 			propertyEditorSchemaAlias: 'Umbraco.MediaPicker3',
+			icon: 'icon-pictures',
+			group: '#propertyEditorUIGroups_media',
+			keywords: [...keywords, 'multiple', 'several'],
+			supportsReadOnly: true,
+		},
+	},
+	{
+		type: 'propertyEditorUi',
+		alias: 'Umb.PropertyEditorUi.SingleMediaPicker',
+		name: 'Single Media Picker Property Editor UI',
+		element: () => import('./property-editor-ui-single-media-picker.element.js'),
+		meta: {
+			label: 'Single Media Picker',
+			propertyEditorSchemaAlias: 'Umbraco.SingleMediaPicker',
 			icon: 'icon-picture',
 			group: '#propertyEditorUIGroups_media',
-			keywords: [
-				'select',
-				'image',
-				'photo',
-				'picture',
-				'banner',
-				'thumbnail',
-				'logo',
-				'avatar',
-				'gallery',
-				'media',
-				'video',
-				'file',
-				'attachment',
-				'cover',
-			],
+			keywords: [...keywords, 'single', 'one'],
 			supportsReadOnly: true,
 		},
 	},
 	schemaManifest,
+	singleSchemaManifest,
 	...valueSummaryManifests,
 ];
