+++ src/Umbraco.Web.UI.Client/src/packages/media/media/property-editors/media-picker/property-editor-ui-media-picker.element.ts
@@ -1,164 +1,12 @@
-import type { UmbInputRichMediaElement } from '../../components/input-rich-media/input-rich-media.element.js';
-import type { UmbCropModel, UmbMediaPickerValueModel } from '../types.js';
-import { UMB_MEDIA_ENTITY_TYPE } from '../../entity.js';
-import { customElement, html, property, state } from '@umbraco-cms/backoffice/external/lit';
-import { UmbChangeEvent } from '@umbraco-cms/backoffice/event';
-import { UmbLitElement } from '@umbraco-cms/backoffice/lit-element';
-import { UmbPropertyEditorUiInteractionMemoryManager } from '@umbraco-cms/backoffice/property-editor';
-import { UMB_PROPERTY_CONTEXT } from '@umbraco-cms/backoffice/property';
-import { UMB_VALIDATION_EMPTY_LOCALIZATION_KEY, UmbFormControlMixin } from '@umbraco-cms/backoffice/validation';
-import type { UmbInteractionMemoryModel } from '@umbraco-cms/backoffice/interaction-memory';
-import type {
-	UmbPropertyEditorConfigCollection,
-	UmbPropertyEditorUiElement,
-} from '@umbraco-cms/backoffice/property-editor';
-import type { UmbNumberRangeValueType } from '@umbraco-cms/backoffice/models';
-import type { UmbTreeStartNode } from '@umbraco-cms/backoffice/tree';
+import { UmbPropertyEditorUIMediaPickerElementBase } from './property-editor-ui-media-picker-base.element.js';
+import { customElement } from '@umbraco-cms/backoffice/external/lit';
 
-import '../../components/input-rich-media/input-rich-media.element.js';
 /**
  * @element umb-property-editor-ui-media-picker
  */
 @customElement('umb-property-editor-ui-media-picker')
-export class UmbPropertyEditorUIMediaPickerElement
-	extends UmbFormControlMixin<UmbMediaPickerValueModel | undefined, typeof UmbLitElement, undefined>(UmbLitElement)
-	implements UmbPropertyEditorUiElement
-{
-	public set config(config: UmbPropertyEditorConfigCollection | undefined) {
-		this.#interactionMemoryManager.setPropertyEditorConfig(config);
-
-		if (!config) return;
-
-		this._allowedMediaTypes = config.getValueByAlias<string>('filter')?.split(',') ?? undefined;
-		this._focalPointEnabled = Boolean(config.getValueByAlias('enableLocalFocalPoint'));
-		this._multiple = Boolean(config.getValueByAlias('multiple'));
-		this._preselectedCrops = config?.getValueByAlias<Array<UmbCropModel>>('crops') ?? [];
-
-		const startNodeId = config.getValueByAlias<string>('startNodeId') ?? '';
-		this._startNode = startNodeId ? { unique: startNodeId, entityType: UMB_MEDIA_ENTITY_TYPE } : undefined;
-
-		const minMax = config.getValueByAlias<UmbNumberRangeValueType>('validationLimit');
-		this._min = minMax?.min ?? 0;
-		this._max = minMax?.max ?? Infinity;
-	}
-
-	/**
-	 * Sets the input to mandatory, meaning validation will fail if the value is empty.
-	 * @type {boolean}
-	 */
-	@property({ type: Boolean })
-	mandatory?: boolean;
-
-	@property({ type: String })
-	mandatoryMessage = UMB_VALIDATION_EMPTY_LOCALIZATION_KEY;
-
-	/**
-	 * Sets the input to readonly mode, meaning value cannot be changed but still able to read and select its content.
-	 * @type {boolean}
-	 * @attr
-	 * @default false
-	 */
-	@property({ type: Boolean, reflect: true })
-	readonly = false;
-
-	@state()
-	private _startNode?: UmbTreeStartNode;
-
-	@state()
-	private _focalPointEnabled: boolean = false;
-
-	@state()
-	private _preselectedCrops: Array<UmbCropModel> = [];
-
-	@state()
-	private _allowedMediaTypes?: Array<string>;
-
-	@state()
-	private _multiple: boolean = false;
-
-	@state()
-	private _min: number = 0;
-
-	@state()
-	private _max: number = Infinity;
-
-	@state()
-	private _alias?: string;
-
-	@state()
-	private _variantId?: string;
-
-	@state()
-	private _interactionMemories: Array<UmbInteractionMemoryModel> = [];
-
-	#interactionMemoryManager = new UmbPropertyEditorUiInteractionMemoryManager(this, {
-		memoryUniquePrefix: 'UmbMediaPicker',
-	});
-
-	constructor() {
-		super();
-
-		this.consumeContext(UMB_PROPERTY_CONTEXT, (context) => {
-			this.observe(context?.alias, (alias) => (this._alias = alias));
-			this.observe(context?.variantId, (variantId) => (this._variantId = variantId?.toString() || 'invariant'));
-		});
-
-		this.observe(
-			this.#interactionMemoryManager.memoriesForPropertyEditor,
-			(interactionMemories) => {
-				this._interactionMemories = interactionMemories ?? [];
-			},
-			null,
-		);
-	}
-
-	override firstUpdated() {
-		this.addFormControlElement(this.shadowRoot!.querySelector('umb-input-rich-media')!);
-	}
-
-	override focus() {
-		return this.shadowRoot?.querySelector<UmbInputRichMediaElement>('umb-input-rich-media')?.focus();
-	}
-
-	#onChange(event: CustomEvent & { target: UmbInputRichMediaElement }) {
-		const isEmpty = event.target.value?.length === 0;
-		this.value = isEmpty ? undefined : event.target.value;
-		this.dispatchEvent(new UmbChangeEvent());
-	}
-
-	async #onInputInteractionMemoriesChange(event: UmbChangeEvent) {
-		const target = event.target as UmbInputRichMediaElement;
-		const interactionMemories = target.interactionMemories;
-
-		if (interactionMemories && interactionMemories.length > 0) {
-			await this.#interactionMemoryManager.saveMemoriesForPropertyEditor(interactionMemories);
-		} else {
-			await this.#interactionMemoryManager.deleteMemoriesForPropertyEditor();
-		}
-	}
-
-	override render() {
-		return html`
-			<umb-input-rich-media
-				.alias=${this._alias}
-				.allowedContentTypeIds=${this._allowedMediaTypes}
-				.focalPointEnabled=${this._focalPointEnabled}
-				.value=${this.value ?? []}
-				.max=${this._max}
-				.min=${this._min}
-				.preselectedCrops=${this._preselectedCrops}
-				.startNode=${this._startNode}
-				.variantId=${this._variantId}
-				.required=${this.mandatory}
-				.requiredMessage=${this.mandatoryMessage}
-				?multiple=${this._multiple}
-				@change=${this.#onChange}
-				?readonly=${this.readonly}
-				.interactionMemories=${this._interactionMemories}
-				@interaction-memories-change=${this.#onInputInteractionMemoriesChange}>
-			</umb-input-rich-media>
-		`;
-	}
+export class UmbPropertyEditorUIMediaPickerElement extends UmbPropertyEditorUIMediaPickerElementBase {
+	protected override readonly multiple = true;
 }
 
 export { UmbPropertyEditorUIMediaPickerElement as element };
